combine door and lift plugin files
